# cpbstats-api/src/main/webapp/api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "CPB Stats APIs"
  },
  "basePath" : "/cpbstats-api/rest",
  "tags" : [ {
    "name" : "Exercices"
  }, {
    "name" : "Authentication"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/authentication/login" : {
      "get" : {
        "tags" : [ "Authentication" ],
        "summary" : "Login",
        "description" : "",
        "operationId" : "login",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "Username",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "User's password",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserGroupApiModel"
            }
          },
          "403" : {
            "description" : "Forbidden"
          }
        }
      }
    },
    "/authentication/logout" : {
      "get" : {
        "tags" : [ "Authentication" ],
        "summary" : "Logout",
        "description" : "",
        "operationId" : "logout",
        "responses" : {
          "403" : {
            "description" : "Forbidden"
          }
        }
      }
    },
    "/authentication/status" : {
      "get" : {
        "tags" : [ "Authentication" ],
        "summary" : "Statut",
        "description" : "",
        "operationId" : "status",
        "responses" : {
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Error"
          }
        }
      }
    },
    "/exercices" : {
      "get" : {
        "tags" : [ "Exercices" ],
        "summary" : "Liste des exercices de l'utilisateur authentifié",
        "description" : "",
        "operationId" : "getExercices",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Exercices" ],
        "summary" : "Eregistre un exercice",
        "description" : "",
        "operationId" : "addExercices",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Exercice à sauvegarder",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Exercice"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/exercices/types" : {
      "get" : {
        "tags" : [ "Exercices" ],
        "summary" : "Liste des types d'exercices",
        "description" : "",
        "operationId" : "getTypeExercices",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/exercices/{id}" : {
      "delete" : {
        "tags" : [ "Exercices" ],
        "summary" : "Liste des exercices de l'utilisateur authentifié",
        "description" : "",
        "operationId" : "deleteExercices",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "Exercice" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Identifiant de l'exercice"
        },
        "typeExerciceId" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Identifiant du type d'exercice"
        },
        "data" : {
          "type" : "string",
          "description" : "Donnee"
        },
        "date" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date"
        }
      }
    },
    "UserGroupApiModel" : {
      "properties" : {
        "username" : {
          "type" : "string"
        },
        "groups" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    }
  }
}